Buttons now update color on press and arrival
Added a callback function to the ElevatorControl class that gets called
on each received request and each floor arrival with the floor number,
travel direction if relevant, and source of the callback. Uncovered a
bug in how turnarounds are handled: when the car moves up to receive a
passenger who wants to go down, it doesn't handle the job queue
correctly.