Moved button control logic, addressed bug
Addressed bug mentioned in the previous commit, where the elevator would
pick up passengers going down when they requested the elevator before
another request to go further up. (E.g. elevator starts on 1. First
request is to go down from 4, second request is to go to 6. Spec says
the elevator should continue to 6 first, then return back to 4 to go
down. This wasn't happening.)

In the process of fixing this, the button control logic got very large,
so it was moved into ElevatorControl and a simpler state updater was
left in ElevatorDisplay. Also added "events" as a dependency, for an
emitter needed to abort elevator requests when they don't line up with
the current direction.