Fix bug where buttons weren't turning off
When a floor button was pressed while the elevator was already at the
floor, if the direction pressed didn't line up with the elevator's
current direction, sometimes it would send the wrong callback to turn
off the button. This was because the way travel direction was calculated
at stops was not accurate.